move to BubbleDate component
