+++ apps/meteor/client/views/room/BubbleDate/index.tsx
@@ -1,0 +1,1 @@
+export * from './BubbleDate';
