+++ apps/meteor/client/views/room/contextualBar/Threads/components/ThreadMessageList.tsx
@@ -1,5 +1,5 @@
 import type { IMessage, IThreadMainMessage } from '@rocket.chat/core-typings';
-import { Box, Bubble } from '@rocket.chat/fuselage';
+import { Box } from '@rocket.chat/fuselage';
 import { useMergedRefs } from '@rocket.chat/fuselage-hooks';
 import { useSetting, useUserPreference } from '@rocket.chat/ui-contexts';
 import { differenceInSeconds } from 'date-fns';
@@ -9,7 +9,7 @@ import React, { Fragment } from 'react';
 import { MessageTypes } from '../../../../../../app/ui-utils/client';
 import { isTruthy } from '../../../../../../lib/isTruthy';
 import { CustomScrollbars } from '../../../../../components/CustomScrollbars';
-import { useFormatDate } from '../../../../../hooks/useFormatDate';
+import { BubbleDate } from '../../../BubbleDate';
 import { isMessageNewDay } from '../../../MessageList/lib/isMessageNewDay';
 import MessageListProvider from '../../../MessageList/providers/MessageListProvider';
 import LoadingMessagesIndicator from '../../../body/LoadingMessagesIndicator';
@@ -49,16 +49,7 @@ type ThreadMessageListProps = {
 };
 
 const ThreadMessageList = ({ mainMessage }: ThreadMessageListProps): ReactElement => {
-	const formatDate = useFormatDate();
-	const {
-		innerRef,
-		bubbleRef,
-		listStyle,
-		bubbleDate,
-		showBubble,
-		style: bubbleDateStyle,
-		className: bubbleDateClassName,
-	} = useDateScroll();
+	const { innerRef, bubbleRef, listStyle, ...bubbleDate } = useDateScroll();
 
 	const { messages, loading } = useLegacyThreadMessages(mainMessage._id);
 	const {
@@ -80,13 +71,7 @@ const ThreadMessageList = ({ mainMessage }: ThreadMessageListProps): ReactElemen
 
 	return (
 		<div className={['thread-list js-scroll-thread', hideUsernames && 'hide-usernames'].filter(isTruthy).join(' ')}>
-			<Box ref={bubbleRef} className={[bubbleDateClassName, showBubble && 'bubble-visible']} style={bubbleDateStyle}>
-				{bubbleDate && (
-					<Bubble small secondary>
-						{formatDate(bubbleDate)}
-					</Bubble>
-				)}
-			</Box>
+			<BubbleDate ref={bubbleRef} {...bubbleDate} />
 			<CustomScrollbars
 				ref={scrollRef}
 				onScroll={(args) => {
